requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
bcrypt==4.0.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.104.1
fastapi-cache2==0.2.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
flower==2.0.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
humanize==4.9.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
makefun==1.15.2
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pendulum==2.1.2
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.23.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
yarl==1.9.3


